refactor search feature modules and update namespaces
