+++ app/src/main/java/com/paris_2/aflami/AflamiApp.kt
@@ -1,9 +1,7 @@
 package com.paris_2.aflami
 
 import android.app.Application
-import com.example.search.di.SearchModule
 import com.paris_2.aflami.di.FireBaseModule
-import com.paris_2.aflami.di.NetworkModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -18,8 +16,6 @@ class AflamiApp : Application() {
             modules(
                 listOf(
                     FireBaseModule,
-                    NetworkModule,
-                    SearchModule
                 )
             )
         }
